:video_game: Menambahkan Controller
Menambahkan Controller sebagai rancangan MVC (Models - Views - Controller)
