+++ app.js
@@ -1,26 +1,27 @@
+// Import Libraries
 const express = require('express');
 const bodyParser = require('body-parser');
 const path = require('path');
 
-// View Enigner
+// Import Routing & Controller
+const shopRoute = require('./routes/shop');
+const adminRoute = require('./routes/admin');
+const errorController = require('./controllers/error');
+
+// View Engine
 const app = express();
 app.set('view engine','ejs');
 app.set('views','views');
-
-const shopRoute = require('./routes/shop');
-const adminRoute = require('./routes/admin').adminRoutes;
-
 app.use(bodyParser.urlencoded({extended:false}));
 app.use(express.static(path.join(__dirname,'public')));
 
+// Page Routing
 app.use(shopRoute);
 app.use('/admin',adminRoute);
 
-app.use((req,res,next)=>{
-    res.render('404',{
-        'title':'Nusantaran JS | Page Not Found'
-    });
-});
+// Handling Error
+app.use(errorController.get404);
 
+// Running Server
 console.log(`Server running at http://127.0.0.1:3000`);
 app.listen(3000);
